+++ torch_frame/data/dataset.py
@@ -60,26 +60,28 @@ def requires_post_materialization(func):
     return _requires_post_materialization
 
 
-def canonicalize_col_to_pattern(col_to_pattern: Union[Optional[str],
-                                                      Dict[str, str]],
-                                columns: List[str]) -> Dict[str, str]:
+def canonicalize_col_to_pattern(col_to_pattern: Union[Optional[Any],
+                                                      Dict[str, Any]],
+                                columns: List[str]) -> Dict[str, Any]:
     r"""Canonicalize :obj:`col_to_pattern` into a dictionary format.
 
     Args:
-        col_to_pattern (Union[str, Dict[str, str]]): A dictionary or a string
-            specifying the separator/pattern for the multi-categorical
-            or timestamp columns. If a string is specified, then the same
-            separator/format will be used throughout all the multi-categorical
-            or timestamp columns. If a dictionary is given, we use a separator
-            specified for each column. (default: :obj:`,`)
-        columns (List[str]): A list of multi-categorical or timestamp columns.
+        col_to_pattern (Union[Any, Dict[str, Any]]): A dictionary or an object
+            specifying the separator/pattern/configuration for the
+            multi-categorical, timestamp or text columns. If an object is
+            specified, then the same separator/pattern/configuration will be
+            used throughout all the multi-categorical, timestamp and text
+            columns. If a dictionary is given, we use each item in the
+            dictionary specified for each column.
+        columns (List[str]): A list of multi-categorical, timestamp or text
+            columns.
 
     Returns:
-        Dict[str, str]: :obj:`col_to_pattern` in a dictionary format, mapping
-            multi-categorical or timestamp columns into their specified
-            separators.
+        Dict[str, Any]: :obj:`col_to_pattern` in a dictionary format, mapping
+            multi-categorical, timestamp or text columns into their specified
+            separators, patterns or configurations.
     """
-    if col_to_pattern is None or isinstance(col_to_pattern, str):
+    if col_to_pattern is None or not isinstance(col_to_pattern, dict):
         pattern = col_to_pattern
         col_to_pattern = {}
         for col in columns:
@@ -111,12 +113,12 @@ class DataFrameToTensorFrameConverter:
             be used throughout all the multi-categorical columns. If a
             dictionary is given, we use a separator specified for each
             column. (default: :obj:`,`)
-        text_embedder_cfg
-            (:class:`torch_frame.config.TextEmbedderConfig`, optional):
-            A text embedder config specifying :obj:`text_embedder` that
-            maps sentences into :class:`torch.nn.Embeddings` and
-            :obj:`batch_size` that specifies the mini-batch size for
-            :obj:`text_embedder`. (default: :obj:`None`)
+        col_to_text_embedder_cfg (TextEmbedderConfig or dict, optional):
+            A text embedder configuration or a dictionary of configurations
+            specifying :obj:`text_embedder` that maps text columns into
+            :class:`torch.nn.Embeddings` and :obj:`batch_size` that
+            specifies the mini-batch size for :obj:`text_embedder`.
+            (default: :obj:`None`)
         text_tokenizer_cfg
             (:class:`torch_frame.config.TextTokenizerConfig`, optional):
             A text tokenizer config specifying :obj:`text_tokenizer` that
@@ -133,7 +135,7 @@ class DataFrameToTensorFrameConverter:
             for more information on formats. If a string is specified,
             then the same format will be used throughout all the timestamp
             columns. If a dictionary is given, we use a different format
-            specified for each column. If not specified, pandas's internal
+            specified for each column. If not specified, Pandas' internal
             to_datetime function will be used to auto parse time columns.
             (default: None)
     """
@@ -143,14 +145,14 @@ class DataFrameToTensorFrameConverter:
         col_stats: Dict[str, Dict[StatType, Any]],
         target_col: Optional[str] = None,
         col_to_sep: Union[str, Dict[str, str]] = ",",
-        text_embedder_cfg: Optional[TextEmbedderConfig] = None,
+        col_to_text_embedder_cfg: Optional[Union[Dict[str, TextEmbedderConfig],
+                                                 TextEmbedderConfig]] = None,
         text_tokenizer_cfg: Optional[TextTokenizerConfig] = None,
         col_to_time_format: Optional[Union[str, Dict[str, str]]] = None,
     ):
         self.col_to_stype = col_to_stype
         self.col_stats = col_stats
         self.target_col = target_col
-        self.text_embedder_cfg = text_embedder_cfg
         self.text_tokenizer_cfg = text_tokenizer_cfg
 
         # Pre-compute a canonical `col_names_dict` for tensor frame.
@@ -175,10 +177,16 @@ class DataFrameToTensorFrameConverter:
             self.col_names_dict.get(torch_frame.timestamp, []),
         )
 
-        if (torch_frame.text_embedded
-                in self.col_names_dict) and (self.text_embedder_cfg is None):
-            raise ValueError("`text_embedder_cfg` needs to be specified when "
-                             "stype.text_embedded column exists.")
+        self.col_to_text_embedder_cfg = canonicalize_col_to_pattern(
+            col_to_text_embedder_cfg,
+            self.col_names_dict.get(torch_frame.text_embedded, []),
+        )
+
+        if (torch_frame.text_embedded in self.col_names_dict) \
+                and (self.col_to_text_embedder_cfg is None):
+            raise ValueError("`col_to_text_embedder_cfg` needs to be"
+                             "specified when stype.text_embedded column "
+                             "exists.")
 
         if (torch_frame.text_tokenized
                 in self.col_names_dict) and (self.text_tokenizer_cfg is None):
@@ -205,9 +213,10 @@ class DataFrameToTensorFrameConverter:
             return TimestampTensorMapper(
                 format=self.col_to_time_format.get(col, None))
         elif stype == torch_frame.text_embedded:
+            text_embedder_cfg = self.col_to_text_embedder_cfg[col]
             return TextEmbeddingTensorMapper(
-                self.text_embedder_cfg.text_embedder,
-                self.text_embedder_cfg.batch_size,
+                text_embedder_cfg.text_embedder,
+                text_embedder_cfg.batch_size,
             )
         elif stype == torch_frame.text_tokenized:
             return TextTokenizationTensorMapper(
@@ -278,9 +287,12 @@ class Dataset(ABC):
             dictionary is given, we use a separator specified for each
             column. (default: :obj:`,`)
             (default: :obj:`,`)
-        text_embedder_cfg (TextEmbedderConfig, optional): A text embedder
-            configuration that specifies the text embedder to map text columns
-            into :pytorch:`PyTorch` embeddings. (default: :obj:`None`)
+        col_to_text_embedder_cfg (TextEmbedderConfig or dict, optional):
+            A text embedder configuration or a dictionary of configurations
+            specifying :obj:`text_embedder` that maps text columns into
+            :class:`torch.nn.Embeddings` and :obj:`batch_size` that
+            specifies the mini-batch size for :obj:`text_embedder`.
+            (default: :obj:`None`)
         text_tokenizer_cfg (TextTokenizerConfig, optional):
             A text tokenizer config specifying :obj:`text_tokenizer` that
             maps sentences into a list of dictionary of tensors. Each
@@ -307,7 +319,8 @@ class Dataset(ABC):
         target_col: Optional[str] = None,
         split_col: Optional[str] = None,
         col_to_sep: Union[str, Dict[str, str]] = ",",
-        text_embedder_cfg: Optional[TextEmbedderConfig] = None,
+        col_to_text_embedder_cfg: Optional[Union[Dict[str, TextEmbedderConfig],
+                                                 TextEmbedderConfig]] = None,
         text_tokenizer_cfg: Optional[TextTokenizerConfig] = None,
         col_to_time_format: Optional[Union[str, Dict[str, str]]] = None,
     ):
@@ -341,7 +354,7 @@ class Dataset(ABC):
             raise ValueError(
                 "Multilabel classification task is not yet supported.")
 
-        self.text_embedder_cfg = text_embedder_cfg
+        self.col_to_text_embedder_cfg = col_to_text_embedder_cfg
         self.text_tokenizer_cfg = text_tokenizer_cfg
         self.col_to_sep = canonicalize_col_to_pattern(
             col_to_sep,
@@ -355,6 +368,13 @@ class Dataset(ABC):
             [
                 col for col, stype in self.col_to_stype.items()
                 if stype == torch_frame.timestamp
+            ],
+        )
+        self.col_to_text_embedder_cfg = canonicalize_col_to_pattern(
+            col_to_text_embedder_cfg,
+            [
+                col for col, stype in self.col_to_stype.items()
+                if stype == torch_frame.text_embedded
             ],
         )
         self._is_materialized: bool = False
@@ -515,7 +535,7 @@ class Dataset(ABC):
             col_stats=self._col_stats,
             target_col=self.target_col,
             col_to_sep=self.col_to_sep,
-            text_embedder_cfg=self.text_embedder_cfg,
+            col_to_text_embedder_cfg=self.col_to_text_embedder_cfg,
             text_tokenizer_cfg=self.text_tokenizer_cfg,
             col_to_time_format=self.col_to_time_format,
         )
@@ -525,14 +545,13 @@ class Dataset(ABC):
         if torch_frame.text_embedded in self._tensor_frame.feat_dict:
             # Text embedding dimensionality is only available after the tensor
             # frame actually gets created, so we compute col_stats here.
-            # For now, we set all EMB_DIM to be the same.
-            # TODO: Extend this to allow different embedding dimensionalities
-            # for different columns.
-            emb_dim = self._tensor_frame.feat_dict[
-                torch_frame.text_embedded].size(-1)
-            for col_name in self._tensor_frame.col_names_dict[
-                    torch_frame.text_embedded]:
-                self._col_stats[col_name][StatType.EMB_DIM] = emb_dim
+            offset = self._tensor_frame.feat_dict[
+                torch_frame.text_embedded].offset
+            emb_dim_list = offset[1:] - offset[:-1]
+            for i, col_name in enumerate(self._tensor_frame.col_names_dict[
+                    torch_frame.text_embedded]):
+                self._col_stats[col_name][StatType.EMB_DIM] = int(
+                    emb_dim_list[i])
 
     @property
     def is_materialized(self) -> bool: