+++ torch_frame/datasets/fake.py
@@ -1,7 +1,7 @@
 import random
 import string
 from datetime import datetime, timedelta
-from typing import List, Optional
+from typing import Dict, List, Optional, Union
 
 import numpy as np
 import pandas as pd
@@ -47,13 +47,6 @@ class FakeDataset(torch_frame.data.Dataset):
         create_split (bool): Whether to create a train, val and test
                 split for the fake dataset. (default: :obj:`False`)
         task_type (TaskType): Task type (default: :obj:`TaskType.REGRESSION`)
-        text_embedder_cfg (TextEmbedderConfig, optional): A text embedder
-            config specifying :obj:`text_embedder` that maps sentences into
-            PyTorch embeddings and :obj:`batch_size` that specifies the
-            mini-batch size for :obj:`text_embedder` (default: :obj:`None`)
-        text_tokenizer_cfg (TextTokenizerConfig, optional): A text tokenizer
-            configuration that specifies the text tokenizer to map text columns
-            into maps sentences into tensor of tokens (default: :obj:`None`)
     """
     def __init__(
         self,
@@ -62,7 +55,8 @@ class FakeDataset(torch_frame.data.Dataset):
         stypes: List[stype] = [stype.categorical, stype.numerical],
         create_split: bool = False,
         task_type: TaskType = TaskType.REGRESSION,
-        text_embedder_cfg: Optional[TextEmbedderConfig] = None,
+        col_to_text_embedder_cfg: Optional[Union[Dict[str, TextEmbedderConfig],
+                                                 TextEmbedderConfig]] = None,
         text_tokenizer_cfg: Optional[TextTokenizerConfig] = None,
     ):
         assert len(stypes) > 0
@@ -74,8 +68,13 @@ class FakeDataset(torch_frame.data.Dataset):
             col_to_stype = {'target': stype.numerical}
         elif task_type == TaskType.MULTICLASS_CLASSIFICATION:
             labels = np.random.randint(0, 3, size=(num_rows, ))
-            # make sure the final label exists
-            labels[0] = 2
+            if num_rows < 3:
+                raise ValueError("Number of rows needs to be at"
+                                 " least 3 for multiclass classification")
+            # make sure every label exists
+            labels[0] = 0
+            labels[1] = 1
+            labels[2] = 2
             df_dict = {'target': labels}
             col_to_stype = {'target': stype.categorical}
         elif task_type == TaskType.BINARY_CLASSIFICATION:
@@ -196,7 +195,7 @@ class FakeDataset(torch_frame.data.Dataset):
             col_to_stype,
             target_col='target',
             split_col='split' if create_split else None,
-            text_embedder_cfg=text_embedder_cfg,
+            col_to_text_embedder_cfg=col_to_text_embedder_cfg,
             text_tokenizer_cfg=text_tokenizer_cfg,
             col_to_time_format={
                 f'timestamp_{i}': TIME_FORMATS[i]