+++ torch_frame/nn/__init__.py
@@ -1,6 +1,7 @@
 r"""Neural network module package."""
 from .base import Module
 from .encoder import *  # noqa
+from .encoding import *  # noqa
 from .conv import *  # noqa
 from .decoder import *  # noqa
 from .models import *  # noqa