+++ torch_frame/utils/infer_stype.py
@@ -57,6 +57,9 @@ def infer_series_stype(ser: Series) -> Optional[stype]:
     if has_nan:
         ser = ser.dropna()
 
+    if len(ser) == 0:
+        return None
+
     # Categorical minimum counting threshold. If the count of the most minor
     # categories is larger than this value, we treat the column as categorical.
     cat_min_count_thresh = 4
@@ -144,5 +147,7 @@ def infer_df_stype(df: DataFrame) -> Dict[str, stype]:
     """
     col_to_stype = {}
     for col in df.columns:
-        col_to_stype[col] = infer_series_stype(df[col])
+        stype = infer_series_stype(df[col])
+        if stype is not None:
+            col_to_stype[col] = stype
     return col_to_stype